refactor(sink): P0 — collapse AcceptedSinkAdapter to AcceptedWriteRequest context
- Move AcceptedWriteRequest struct from write/accepted.rs into format.rs,
  co-located with the trait it belongs to
- Rename AcceptedSinkAdapter::write_accepted(9 args) → write(req)
- Remove #[allow(clippy::too_many_arguments)]
- Remove AdapterAcceptedWriter wrapper and AcceptedWriter trait (both
  now redundant); write_with_adapter delegates to adapter.write() directly
- Delta, Parquet, Iceberg impls destructure req at the top of write()
